перенес тело функции в метод класса
